requirements.txt
ultralytics
opencv-python
asyncio
numpy
fastapi
logging
uvicorn[standard]
pyyaml
